Refactor WASM helper for tests.
